move the `enhance` to the `agent`
